pyproject.toml
[tool.poetry]
name = "EDI-Tool"
version = "0.1.0"
description = ""
authors = ["maxomm <max.warsinke@googlemail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
matplotlib = "^3.5.0"
Pillow = "^9.0.0"
pytest = "^7.1.2"
keras = "2.9.0"
tensorflow = "2.9.0"
auto-py-to-exe = "^2.22.0"
opencv-python = "^4.6.0.66"
scipy = "^1.10.0"

[tool.poetry.dev-dependencies]
bandit = "^1.7.2"
black = "^22.1.0"
flake8 = "^4.0.1"
flake8-black = "^0.2.4"
flake8-bugbear = "^22.1.11"
flake8-builtins = "^1.5.3"
flake8-isort = "^4.1.1"
pep8-naming = "^0.12.1"
pylint = "^2.12.2"
pytest = "^7.0.1"
isort = "^5.10.1"
mypy = "^0.931"
rope = "^0.22.0"
vulture = "^2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
known_first_party = "MYPACKAGENAME"
known_local_folder = "tests"